allow full scrub fields intake from request

+++ tests/client/aiohttp/test_async_requests.py
@@ -4,6 +4,7 @@ import time
 
 import pytest
 
+import quickbolt.reporting.response_csv as rc
 from quickbolt.clients import AioRequests
 
 pytestmark = pytest.mark.client
@@ -13,12 +14,16 @@ headers = {}
 url = "https://jsonplaceholder.typicode.com/users/1"
 
 
-def test_request(batch=None, delay=0, report=True, **kwargs):
+def test_request(
+    batch=None, delay=0, report=True, full_scrub_fields=None, delete=True, **kwargs
+):
     pytest.aio_requests = AioRequests(root_dir=root_dir)
     pytest.run_info_path = pytest.aio_requests.logging.run_info_path
 
     batch = batch or {"method": "get", "headers": headers, "url": url}
-    response = pytest.aio_requests.request(batch, delay=delay, report=report, **kwargs)
+    response = pytest.aio_requests.request(
+        batch, delay=delay, report=report, full_scrub_fields=full_scrub_fields, **kwargs
+    )
 
     assert response.get("duration")
 
@@ -52,9 +57,10 @@ def test_request(batch=None, delay=0, report=True, **kwargs):
     else:
         assert responses[0].get("message")
 
-    asyncio.run(pytest.aio_requests.logging.delete_run_info(root_dir))
-    path = pytest.aio_requests.logging.log_file_path
-    assert not os.path.exists(path)
+    if delete:
+        asyncio.run(pytest.aio_requests.logging.delete_run_info(root_dir))
+        path = pytest.aio_requests.logging.log_file_path
+        assert not os.path.exists(path)
 
 
 def test_request_multiple():
@@ -80,3 +86,36 @@ def test_request_content_stream():
 def test_request_report():
     test_request(report=False)
     assert not os.path.exists(pytest.aio_requests.csv_path)
+
+
+def test_request_report_full_scrub_fields():
+    full_scrub_fields = ["message"]
+    test_request(full_scrub_fields=full_scrub_fields, delete=False)
+    expected_message = {
+        "id": "0000000",
+        "name": "0000000000000",
+        "username": "0000",
+        "email": "00000000000000000",
+        "address": {
+            "street": "00000000000",
+            "suite": "00000000",
+            "city": "00000000000",
+            "zipcode": "0000000000",
+            "geo": {"lat": "00000000", "lng": "0000000"},
+        },
+        "phone": "000000000000000000000",
+        "website": "0000000000000",
+        "company": {
+            "name": "000000000000000",
+            "catchPhrase": "00000000000000000000000000000000000000",
+            "bs": "000000000000000000000000000",
+        },
+    }
+
+    scrubbed_csv_path = pytest.aio_requests.csv_path.replace(".csv", "_scrubbed.csv")
+    scrubbed_dict = asyncio.run(rc.csv_to_dict(scrubbed_csv_path))
+    assert scrubbed_dict[0]["MESSAGE"] == expected_message
+
+    asyncio.run(pytest.aio_requests.logging.delete_run_info(root_dir))
+    path = pytest.aio_requests.logging.log_file_path
+    assert not os.path.exists(path)
